add MiniMax-M2.5 test
