+++ test/registered/ascend/accuracy/test_npu_minimax_m2_5_w8a8_1p1d_24p_aime2025.py
@@ -18,7 +18,12 @@ register_npu_ci(
 
 PREFILL_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
+    "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
+    "SGLANG_EXTERNAL_MODEL_PACKAGE": "custom_eagle3",
     "STREAMS_PER_DEVICE": "32",
+    "ENABLE_PROFILING": "1",
+    "PROFILING_BS": "30",
+    "PROFILING_step": "8",
     "ASCEND_USE_FIA": "1",
     "HCCL_BUFFSIZE": "2500",
     "DEEP_NORMAL_MODE_USE_INT8_QUANT": "1",
@@ -26,16 +31,22 @@ PREFILL_ENVS = {
     "DEEPEP_NORMAL_LONG_SEQ_ROUND": "64",
     "DEEPEP_NORMAL_LONG_SEQ_PER_ROUND_TOKENS": "2048",
     "DEEPEP_NORMAL_COMBINE_ENABLE_LONG_SEQ": "1",
+    "HCCL_SOCKET_IFNAME": NIC_NAME,
+    "GLOO_SOCKET_IFNAME": NIC_NAME,
 }
 
 DECODE_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
+    "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
+    "SGLANG_EXTERNAL_MODEL_PACKAGE": "custom_eagle3",
     "STREAMS_PER_DEVICE": "32",
     "HCCL_BUFFSIZE": "1600",
     "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "640",
     "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
     "SGLANG_ENABLE_SPEC_V2": "1",
     "SGLANG_NPU_FUSED_MOE_MODE": "2",
+    "HCCL_SOCKET_IFNAME": NIC_NAME,
+    "GLOO_SOCKET_IFNAME": NIC_NAME,
 }
 
 PREFILL_ARGS = [
@@ -89,8 +100,6 @@ PREFILL_ARGS = [
     "--speculative-draft-model-quantization",
     "unquant",
     "--disable-radix-cache",
-    "--external-model-package",
-    "custom_eagle3",
 ]
 
 DECODE_ARGS = [
@@ -104,6 +113,8 @@ DECODE_ARGS = [
     8,
     16,
     24,
+    "--mem-fraction-static",
+    0.6,
     "--attention-backend",
     "ascend",
     "--device",
@@ -144,8 +155,6 @@ DECODE_ARGS = [
     "--speculative-draft-model-quantization",
     "unquant",
     "--disable-radix-cache",
-    "--external-model-package",
-    "custom_eagle3",
 ]
 
 MODEL_CONFIG = {
