+++ test/registered/ascend/performance/test_npu_minimax_m2_5_w8a8_8p_in3k5_out1k5_20ms.py
@@ -1,68 +1,65 @@
 import unittest
 
-from sglang.test.ascend.e2e.test_npu_accuracy_utils import (
-    BENCHMARK_TOOL_DEFAULT,
-    TestAscendAccuracyTestCaseBase,
-)
 from sglang.test.ascend.e2e.test_npu_performance_utils import (
+    AISBENCHMARK_DATASET_DEFAULT,
+    BENCHMARK_TOOL_DEFAULT,
     MINIMAX_M2_5_W8A8_MODEL_PATH,
     MINIMAX_M2_5_EAGLE3_MODEL_PATH,
+    TestAscendPerformanceTestCaseBase,
 )
 from sglang.test.ci.ci_register import register_npu_ci
 
 register_npu_ci(
-    est_time=3600,
+    est_time=1800,
     suite="nightly-16-npu-a3",
     nightly=True,
+    disabled="Currently it is executed by the npu performance workflow.",
 )
 
-MINIMAX_M2_5_128K_PREFIX_ENVS = {
+MINIMAX_M2_5_LOW_LATENCY_ENVS = {
     "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
     "STREAMS_PER_DEVICE": "32",
     "HCCL_SOCKET_IFNAME": "lo",
     "GLOO_SOCKET_IFNAME": "lo",
+    "HCCL_OP_EXPANSION_MODE": "AIV",
     "TASK_QUEUE_ENABLE": "1",
-    "ASCEND_USE_FIA": "1",
-    "HCCL_BUFFSIZE": "1600",
+    "HCCL_BUFFSIZE": "800",
     "SGLANG_SET_CPU_AFFINITY": "1",
     "SGLANG_ENABLE_SPEC_V2": "1",
-    "SGLANG_ENABLE_OVERLAP_PLAN_STREAM": "1",
-    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "640",
-    "DEEPEP_NORMAL_LONG_SEQ_ROUND": "64",
-    "DEEPEP_NORMAL_LONG_SEQ_PER_ROUND_TOKENS": "2048",
-    "DEEPEP_NORMAL_COMBINE_ENABLE_LONG_SEQ": "1",
     "SGLANG_NPU_FUSED_MOE_MODE": "2",
-    "SGLANG_NPU_DEEPEP_USE_FUSED_MOE_DECODE": "1",
-    "SGLANG_NPU_FUSEEP_DECODE_ONLY": "1",
+    "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "204800",
+    "SGLANG_SCHEDULER_DECREASE_PREFILL_IDLE": "1",
+    "SGLANG_PREFILL_DELAYER_MAX_DELAY_PASSES": "200",
 }
 
-MINIMAX_M2_5_128K_PREFIX_OTHER_ARGS = [
+MINIMAX_M2_5_LOW_LATENCY_OTHER_ARGS = [
     "--model-path",
     MINIMAX_M2_5_W8A8_MODEL_PATH,
-    "--host",
-    "127.0.0.1",
-    "--port",
-    32000,
     "--tp-size",
     16,
-    "--dp-size",
-    2,
     "--enable-dp-attention",
-    "--prefill-delayer-max-delay-passes",
-    100,
-    "--enable-prefill-delayer",
+    "--dp-size",
+    16,
+    "--ep-size",
+    16,
     "--mem-fraction-static",
-    0.65,
+    0.75,
     "--max-running-requests",
-    36,
+    512,
+    "--disable-radix-cache",
+    "--prefill-delayer-max-delay-passes",
+    "--enable-prefill-delayer",
     "--chunked-prefill-size",
     -1,
     "--max-prefill-tokens",
-    130000,
+    4096,
     "--cuda-graph-bs",
+    1,
+    7,
     8,
+    9,
+    10,
     16,
-    24,
     "--moe-a2a-backend",
     "ascend_fuseep",
     "--deepep-mode",
@@ -83,30 +80,31 @@ MINIMAX_M2_5_128K_PREFIX_OTHER_ARGS = [
     "unquant",
     "--dtype",
     "bfloat16",
-    "--trust-remote-code",
-    "--tokenizer-worker-num",
-    8,
     "--external-model-package",
     "custom_eagle3",
 ]
 
 
-class TestNPUMiniMaxM2_5_W8A8_8P_AIME2025(TestAscendAccuracyTestCaseBase):
-    """Test NPU accuracy for MiniMax-M2.5-w8a8 8p single node prefix in128k on AIME 2025"""
+class TestNPUMiniMaxM2_5_W8A8_8P_In3k5_Out1k5_LowLatency(TestAscendPerformanceTestCaseBase):
+    """Test NPU performance for MiniMax-M2.5-w8a8 8p single node low latency in3k5 out1k5"""
 
     benchmark_tool = BENCHMARK_TOOL_DEFAULT
+    aisbench_dataset_type = AISBENCHMARK_DATASET_DEFAULT
     model = MINIMAX_M2_5_W8A8_MODEL_PATH
-    other_args = MINIMAX_M2_5_128K_PREFIX_OTHER_ARGS
-    envs = MINIMAX_M2_5_128K_PREFIX_ENVS
-    accuracy = 0.8
-    dataset_type = "aime2025"
-    dataset_name = "aime2025_gen"
-    batch_size = 64
-    max_out_len = 8192
+    other_args = MINIMAX_M2_5_LOW_LATENCY_OTHER_ARGS
+    envs = MINIMAX_M2_5_LOW_LATENCY_ENVS
+    dataset_name = "random"
+    max_concurrency = 30
+    num_prompts = 120
+    input_len = 3500
+    output_len = 1500
+    random_range_ratio = 1
+    tpot = 20
+    output_token_throughput = 3000
 
-    def test_npu_minimax_m2_5_w8a8_8p_aime2025(self):
-        """Run NPU accuracy test for MiniMax-M2.5-w8a8 8p single node prefix in128k on AIME 2025"""
-        self.run_accuracy()
+    def test_npu_minimax_m2_5_w8a8_8p_in3k5_out1k5_low_latency(self):
+        """Run NPU performance test for MiniMax-M2.5-w8a8 low latency"""
+        self.run_throughput()
 
 
 if __name__ == "__main__":
